Drop old render before rerendering
This results in dramatic speedups because jemalloc is able to
re-use the space allocated by the previous render. Alone, this
isn't such a big deal, but it means that, more often than not,
listeners are re-allocated in the same location. This prevents
having to detatch the old listener location from the DOM and
re-attach the new one.